requirements.txt
amqp==5.3.1
asgiref==3.8.1
billiard==4.2.1
celery==5.4.0
click==8.1.8
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
colorama==0.4.6
cron-descriptor==1.4.5
Django==5.1.4
django-celery-beat==2.7.0
django-js-asset==3.0.1
django-model-utils==5.0.0
django-timezone-field==7.1
jsonfield==3.1.0
kombu==5.4.2
pillow==11.1.0
prompt_toolkit==3.0.48
python-crontab==3.2.0
python-dateutil==2.9.0.post0
pytz==2024.2
redis==5.2.1
setuptools==75.8.0
six==1.17.0
sqlparse==0.5.3
swapper==1.4.0
tzdata==2024.2
vine==5.1.0
wcwidth==0.2.13
